requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
pdf2image>=1.16.0
chromadb>=0.4.0
pydantic>=2.0.0
